+++ src/alited.tcl
@@ -7,7 +7,7 @@
 # License: MIT.
 ###########################################################
 
-package provide alited 1.3.6a2  ;# for documentation (esp. for Ruff!)
+package provide alited 1.3.6b2  ;# for documentation (esp. for Ruff!)
 
 set _ [package require Tk]
 
@@ -361,6 +361,14 @@ namespace eval alited {
   }
   #_______________________
 
+  proc NormalizeName {name} {
+    # Removes spec.characters from a name (sort of normalizing it).
+    #   name - the name
+
+    return [string map [list \\ {} \{ {} \} {} \[ {} \] {} \t {} \n {} \r {} \" {}] $name]
+  }
+  #_______________________
+
   proc NormalizeFileName {name} {
     # Removes spec.characters from a file/dir name (sort of normalizing it).
     #   name - the name of file/dir
@@ -412,6 +420,45 @@ namespace eval alited {
     ::hl_${lng}::hl_init $wtxt -dark [::apave::obj csDark $cs] -colors $colors \
       -multiline 1 -font $al(FONT,txt) -insertwidth $al(CURSORWIDTH) {*}$args
     ::hl_${lng}::hl_text $wtxt
+  }
+  #_______________________
+
+  proc TextIcon {ico {to out}} {
+    # Gets a picture from a character and vice versa.
+    #   ico - picture or character
+    #   to - "in" gets in-chars, "out" gets out-chars
+
+# TODO: codes instead of pictures - not available in 8.6.10
+#!    \U0001f4a5 = 💥
+#!    \U0001f4bb = 💻
+#!    \U0001f3d7 = 🏗
+#!    \U0001f4f6 = 📶
+#!    \U0001f4e1 = 📡
+#!    \U0001f4d6 = 📖
+#!    \U0001f300 = 🌀
+#!    \U0001f4de = 📞
+#!    \U0001f4d0 = 📐
+#!    \U0001f426 = 🐦
+#!    \U0001f381 = 🎁
+#!    \U0001f3c1 = 🏁
+#!    \U0001f511 = 🔑
+#!    \U0001f4be = 💾
+
+    set in {0 1 2 3 4 5 6 7 8 9 & ~ = @}
+    set out {💥 💻 🏗 📶 📡 📖 🌀 📞 📐 🐦 🎁 🏁 🔑 💾}
+    if {$to eq {out}} {
+      set lfrom $in
+      set lto $out
+    } else {
+      set lfrom $out
+      set lto $in
+    }
+    if {[set i [lsearch -exact $lfrom $ico]]>-1} {
+      set tico [lindex $lto $i]
+    } else {
+      set tico $ico
+    }
+    return $tico
   }
 
   ## ________________________ Messages _________________________ ##